Admin component now uses custom comparison function for sorting threads
The (paginated) discussion manager in the admin area so far used the dates of
the available threads as keys in the array used for display preparation. This
lead to a bug with concurrently updated threads (same filemtime) which lead
to all but the last handled thread with the same timestamp becoming
unaccessable via the admin component. A change to a custom comparison
function together with usort instead of krsort fixes this.

darcs-hash:20080401211950-2b4f5-e0dfc62eea3977e3771cd1092626f97c5b835418.gz