fix for discussions not showing when breadcrumbs contain page containing a blog
Rather tough bug. If the breadcrumbs or youarehere traces contain a wiki page
which itself contains a {{blog>}}, when rendering of this page the include
plugin gets loaded. This formerly triggered the _ignore function becoming
active. It is not necessary though, as the discussion plugin renders upon an
event which cannot be triggered by either include or pagelist plugin. Thus
the check was removed and the bug with it.

darcs-hash:20080404160247-2b4f5-b8a263811b9a71b161c7a388e4e38b9c281c244f.gz