Save .comment data only when modified #66 #59
This fixes the problem that .comment files were written whenever the
page was parsed which made the order in the threads syntax and in the
admin interface meaningless.
Old dates will still be meaningless but at least the pages with new
comments should appear in the correct order now though pages will still
move to the top when the comment status is changed. This could be
properly fixed by storing the date of the last comment in the metadata
and indexing the metadata, then it would be easy to get all dates.