Fix discussion status update when the discussion syntax is removed
This changes how the discussion status is saved. First of all the status
is recorded in the metadata of the page and then the status is saved
into the .comments file even when no discussion syntax is in the page.
This makes it possible that the discussion is disabled when the syntax
is removed (unless the "automatic" option is enabled).